IP version independent handling in VIP table, that learn+floods rather than drops traffic to non FAUCET MAC. Only update eth_dst/hairpin rules on port changes.

+++ faucet/dp.py
@@ -305,7 +305,7 @@ configuration.
         test_config_condition(self.timeout < self.arp_neighbor_timeout, (
             'L2 timeout must be >= ARP timeout'))
         if self.cache_update_guard_time == 0:
-            self.cache_update_guard_time = self.timeout / 2
+            self.cache_update_guard_time = int(self.timeout / 2)
         if self.lldp_beacon:
             self._check_conf_types(self.lldp_beacon, self.lldp_beacon_defaults_types)
             test_config_condition('send_interval' not in self.lldp_beacon, (
